# specs/001-note-management-api/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Quick Note — Note Management API
  version: 1.0.0
  summary: Personal note capture, organization, search, and recovery.
  description: |
    This document is the authoritative contract for the Quick Note service. It is hand-authored and
    precedes the implementation, per Constitution Principle I. Where the running service and this
    document disagree, this document is correct and the service is defective.

    **Authentication.** Every operation requires a Bearer JWT issued by the configured Keycloak
    realm. The service verifies signature, `iss`, `aud`, and `exp`, and derives the acting user from
    the `sub` claim. It never issues, refreshes, or revokes tokens. A user identifier supplied
    anywhere else in a request is ignored.

    **Errors.** Every non-2xx response is `application/problem+json` per RFC 9457, carrying a stable
    machine-readable `code`. Codes are part of this contract; see `contracts/error-catalog.md`.

    **Non-disclosure.** A resource that does not exist and a resource owned by another user both
    return `404`. The API never reveals which of the two applies.
servers:
  - url: https://api.quicknote.example/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
  /notes:
    get:
      tags: [Notes]
      operationId: listNotes
      summary: List or search notes
      description: |
        Returns the caller's notes. By default only `ACTIVE` notes are returned; archived and
        trashed notes are excluded (FR-033). Ordering is pinned notes first, then most recently
        updated first (FR-035). Filters, search, and pagination combine (FR-040).
      parameters:
        - $ref: '#/components/parameters/StateFilter'
        - $ref: '#/components/parameters/LabelFilter'
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: A page of notes. Empty when nothing matches — never an error (FR-038).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotePage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    post:
      tags: [Notes]
      operationId: createNote
      summary: Create a note
      description: |
        Creates a note owned by the caller. At least one of `title` or `body` must be present
        (FR-010). Supply `Idempotency-Key` to make retries safe; a replay within 24 hours returns the
        original response.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateNoteRequest' }
      responses:
        '201':
          description: Note created.
          headers:
            Location:
              description: URI of the created note.
              schema: { type: string, format: uri-reference }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Note' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: A referenced label does not exist or is not owned by the caller (FR-032).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /notes/{noteId}:
    parameters:
      - $ref: '#/components/parameters/NoteId'
    get:
      tags: [Notes]
      operationId: getNote
      summary: Retrieve a note
      responses:
        '200':
          description: The note.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Note' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    patch:
      tags: [Notes]
      operationId: updateNote
      summary: Update a note's content or state
      description: |
        Partial update. Only the properties present in the body are changed; an absent property is
        untouched, and an explicit `null` clears `title` or `body`.

        State changes are attribute writes rather than separate verb endpoints. Setting
        `archived: true` clears `pinned` (FR-020). Setting `trashed: false` restores a trashed note
        to the state it held before trashing (FR-023). Any other change to a trashed note is
        rejected `409 note_trashed` (FR-021).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateNoteRequest' }
      responses:
        '200':
          description: Updated note.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Note' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: |
            The note is trashed and must be restored first (`note_trashed`), or a concurrent update
            won (`note_version_conflict`).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    delete:
      tags: [Notes]
      operationId: trashNote
      summary: Move a note to the trash
      description: |
        Moves the note to the trash and clears `pinned` (FR-022). Idempotent: deleting an
        already-trashed note succeeds without changing `trashedAt`. This never destroys data —
        permanent removal is `DELETE /trash/{noteId}`.
      responses:
        '204': { description: Note is in the trash. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /notes/{noteId}/labels:
    parameters:
      - $ref: '#/components/parameters/NoteId'
    put:
      tags: [Notes]
      operationId: replaceNoteLabels
      summary: Replace the note's entire label set
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [labelIds]
              properties:
                labelIds:
                  type: array
                  maxItems: 20
                  uniqueItems: true
                  items: { type: string, format: uuid }
      responses:
        '200':
          description: Updated note.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Note' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: The note is trashed (`note_trashed`).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /notes/{noteId}/labels/{labelId}:
    parameters:
      - $ref: '#/components/parameters/NoteId'
      - $ref: '#/components/parameters/LabelId'
    put:
      tags: [Notes]
      operationId: attachLabel
      summary: Attach a label to a note
      description: Idempotent — attaching an already-attached label succeeds and changes nothing (FR-029).
      responses:
        '204': { description: Label is attached. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: The note or the label does not exist, or is not owned by the caller (FR-032).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
        '409':
          description: The note is trashed (`note_trashed`).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
        '422':
          description: The note already carries 20 labels (`note_label_limit_exceeded`) (FR-030).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    delete:
      tags: [Notes]
      operationId: detachLabel
      summary: Detach a label from a note
      description: Idempotent — detaching an unattached label succeeds and changes nothing (FR-029).
      responses:
        '204': { description: Label is not attached. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: The note is trashed (`note_trashed`).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /labels:
    get:
      tags: [Labels]
      operationId: listLabels
      summary: List the caller's labels
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: A page of labels, ordered by name.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LabelPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    post:
      tags: [Labels]
      operationId: createLabel
      summary: Create a label
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LabelRequest' }
      responses:
        '201':
          description: Label created.
          headers:
            Location:
              schema: { type: string, format: uri-reference }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Label' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/LabelConflict' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /labels/{labelId}:
    parameters:
      - $ref: '#/components/parameters/LabelId'
    patch:
      tags: [Labels]
      operationId: renameLabel
      summary: Rename a label
      description: |
        Renaming changes no note content; every note carrying the label reflects the new name
        immediately (FR-026).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LabelRequest' }
      responses:
        '200':
          description: Renamed label.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Label' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/LabelConflict' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    delete:
      tags: [Labels]
      operationId: deleteLabel
      summary: Delete a label
      description: |
        Permanently deletes the label and detaches it from every note that carried it. No note is
        deleted as a consequence (FR-031). Labels do not go to the trash.
      responses:
        '204': { description: Label deleted. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by the configured Keycloak realm. The service verifies signature, `iss`, `aud`,
        and `exp`, and reads the user identifier from `sub`.
  parameters:
    NoteId:
      name: noteId
      in: path
      required: true
      schema: { type: string, format: uuid }
    LabelId:
      name: labelId
      in: path
      required: true
      schema: { type: string, format: uuid }
    StateFilter:
      name: state
      in: query
      required: false
      description: Which notes to include. Defaults to `active` (FR-033, FR-034).
      schema:
        type: string
        enum: [active, archived, trashed]
        default: active
    LabelFilter:
      name: labelId
      in: query
      required: false
      description: Return only notes carrying this label (FR-034).
      schema: { type: string, format: uuid }
    SearchQuery:
      name: q
      in: query
      required: false
      description: |
        Case-insensitive substring matched against title and body; a note matching in either is
        returned (FR-036). Trashed notes are excluded unless `state=trashed` (FR-037).
      schema: { type: string, minLength: 1, maxLength: 200 }
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      required: false
      description: Page size. Requests above the maximum are rejected, not clamped (FR-039).
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Client-generated key; a replay within 24 hours returns the original response.
      schema: { type: string, minLength: 1, maxLength: 255 }
  schemas:
    NoteColor:
      type: string
      enum: [default, red, orange, yellow, green, teal, blue, dark_blue, purple, pink, brown]
      default: default
    Note:
      type: object
      additionalProperties: false
      required: [id, color, pinned, archived, trashed, labelIds, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        title:
          type: [string, 'null']
          maxLength: 200
        body:
          type: [string, 'null']
          maxLength: 20000
        color: { $ref: '#/components/schemas/NoteColor' }
        pinned:
          type: boolean
          description: Never true while `archived` or `trashed` is true (FR-020).
        archived:
          type: boolean
          description: |
            A note is in exactly one of three states — active, archived, or trashed. `archived` and
            `trashed` are the wire form of that single state and are never both true; both false
            means the note is active.
        trashed:
          type: boolean
          description: See `archived`. Never true at the same time as `archived`.
        trashedAt:
          type: [string, 'null']
          format: date-time
          description: Non-null if and only if `trashed` is true.
        labelIds:
          type: array
          maxItems: 20
          uniqueItems: true
          items: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateNoteRequest:
      type: object
      additionalProperties: false
      description: At least one of `title` or `body` must be present and non-blank (FR-010).
      properties:
        title: { type: [string, 'null'], maxLength: 200 }
        body: { type: [string, 'null'], maxLength: 20000 }
        color: { $ref: '#/components/schemas/NoteColor' }
        pinned: { type: boolean, default: false }
        labelIds:
          type: array
          maxItems: 20
          uniqueItems: true
          items: { type: string, format: uuid }
    UpdateNoteRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      description: |
        Only the properties present are changed. An explicit `null` clears `title` or `body`; the
        result must still satisfy the rule that they are not both empty.
      properties:
        title: { type: [string, 'null'], maxLength: 200 }
        body: { type: [string, 'null'], maxLength: 20000 }
        color: { $ref: '#/components/schemas/NoteColor' }
        pinned: { type: boolean }
        archived:
          type: boolean
          description: Setting true clears `pinned`.
        trashed:
          type: boolean
          description: |
            Setting false restores the note to the state it held before it was trashed. Setting true
            is equivalent to `DELETE /notes/{noteId}`.
    Label:
      type: object
      additionalProperties: false
      required: [id, name, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 50 }
        createdAt: { type: string, format: date-time }
    LabelRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
          description: |
            Trimmed before storage. Unique per user, compared case-insensitively (FR-027). The
            submitted casing is preserved for display.
    NotePage:
      type: object
      additionalProperties: false
      required: [items, page, size, totalElements, totalPages]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Note' }
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1, maximum: 100 }
        totalElements: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
        nextPage:
          type: [string, 'null']
          format: uri-reference
          description: URI of the next page, or null on the last page.
    LabelPage:
      type: object
      additionalProperties: false
      required: [items, page, size, totalElements, totalPages]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Label' }
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1, maximum: 100 }
        totalElements: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
        nextPage:
          type: [string, 'null']
          format: uri-reference
          description: URI of the next page, or null on the last page.
    ProblemDetail:
      type: object
      description: RFC 9457 Problem Details, extended with a stable machine-readable `code`.
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail:
          type: string
          description: Human-readable. Never contains stack traces, SQL, or upstream error text (FR-043).
        instance: { type: string, format: uri-reference }
        code:
          type: string
          description: Stable identifier; part of this contract. See `error-catalog.md`.
        correlationId:
          type: string
          description: Echoes the request correlation ID, for support and log lookup.
        errors:
          type: array
          description: Present on validation failures; one entry per offending field (FR-041).
          items:
            type: object
            required: [field, code, message]
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }
  responses:
    BadRequest:
      description: Malformed request or failed validation.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    Unauthorized:
      description: |
        Missing, expired, malformed, or unverifiable credential. The response does not reveal which
        check failed (FR-003).
      headers:
        WWW-Authenticate:
          schema: { type: string }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    NotFound:
      description: |
        The resource does not exist, or it belongs to another user. These are indistinguishable by
        design (FR-008).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    LabelConflict:
      description: A label with this name already exists for the caller (`label_name_conflict`).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    UnprocessableEntity:
      description: Well-formed but semantically rejected.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    TooManyRequests:
      description: Per-principal or per-IP rate limit exceeded.
      headers:
        Retry-After:
          schema: { type: integer }
        RateLimit-Remaining:
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    PayloadTooLarge:
      description: |
        The request body exceeds the transport byte ceiling. Rejected at the edge, before any handler
        runs, so no partial state is written.

        This is an abuse guard, not the content limit. It sits far above any legitimate note, so a
        well-formed client never sees it: exceeding the 20,000-character `body` or 200-character
        `title` limit returns `400 validation_failed` naming the offending field. Note that those
        limits are counted in **characters**, not bytes — a 20,000-character note in a script whose
        characters encode to four bytes is valid and is accepted.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    InternalError:
      description: |
        An unhandled fault. The body carries only `type`, `title`, `status`, `code`, and
        `correlationId` — never a stack trace, SQL fragment, hostname, or upstream error text
        (FR-043). Full detail is logged against the same correlation ID.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    ServiceUnavailable:
      description: |
        A critical dependency is unavailable. This is also what the readiness probe reflects.
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }